move validation functions to new module
